Split off lexer into multiple sub-modules
The original file was way, WAY too long, really.
